refactor(rig): split god files + dedup + add remote-exec preflight (#5241, #5283, #5285)
#5241 structural: split extension/lifecycle.rs (1962L/31 items) into
lifecycle/{mod,install_sources,update}.rs; split rig/install.rs discovery
group into rig/discovery.rs; extract rig source install/removal from
runner/rig_materialization.rs into a sibling module. All under thresholds;
structural moves only, behavior/contracts preserved.

#5283 dedup: remove_runner_installed_rig_source's duplicated rig-sources exec
block extracted into exec_runner_rig_sources_command helper.

#5285 remote-exec preflight: rig install dispatch now runs
preflight_remote_argv_path_translation (reusing the #5093/#5071 helper) and
passes a RunnerCapabilityPreflight so exec validates remote capability parity
before remote dispatch.

Note: #5241's src/core/defaults.rs (38 items) finding is out of scope for this
rig cluster and deferred.
