+++ src/core/extension/lifecycle/mod.rs
@@ -1,17 +1,9 @@
-use crate::core::config::{self, from_str};
 use crate::core::engine::identifier;
-use crate::core::engine::local_files::{self, FileSystem};
 use crate::core::error::{Error, Result};
-use crate::core::git;
 use crate::core::paths;
-use std::collections::HashMap;
 use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::sync::{Mutex, OnceLock};
 
-use super::execution::run_setup;
-use super::manifest::ExtensionManifest;
-use super::{is_extension_linked, load_extension, ExtensionSourceUpdate};
+use super::load_extension;
 
 #[derive(Debug, Clone)]
 pub struct InstallResult {
@@ -37,7 +29,7 @@ pub struct UpdateResult {
     pub linked: bool,
     pub source_path: Option<PathBuf>,
     pub git_root: Option<PathBuf>,
-    pub source_update: ExtensionSourceUpdate,
+    pub source_update: super::ExtensionSourceUpdate,
     pub repaired_source_metadata: Option<source_metadata::SourceMetadataRepair>,
 }
 
@@ -141,656 +133,17 @@ pub fn install_for_component(
     })
 }
 
-fn install_configured_extension(source: &str, extension_id: &str) -> Result<InstallResult> {
-    if is_git_url(source) {
-        return install(source, Some(extension_id));
-    }
-
-    let source_path = Path::new(source);
-    let candidate = source_path
-        .join(extension_id)
-        .join(format!("{}.json", extension_id));
-
-    if candidate.exists() {
-        let extension_path = source_path.join(extension_id);
-        return install_from_path(
-            &extension_path.to_string_lossy(),
-            Some(extension_id),
-            Some(source_path),
-        );
-    }
-
-    install(source, Some(extension_id))
-}
-
-/// Install a extension by cloning from a git repository URL.
-///
-/// Handles both single-extension repos (manifest at repo root) and monorepos
-/// (manifest in a subdirectory matching the extension ID). For monorepos,
-/// extracts just the target subdirectory.
-fn install_from_url(
-    url: &str,
-    id_override: Option<&str>,
-    revision: Option<&str>,
-) -> Result<InstallResult> {
-    let extension_id = match id_override {
-        Some(id) => slugify_id(id)?,
-        None => derive_id_from_url(url)?,
-    };
-
-    // Check cross-entity name collision before checking extension-specific existence
-    config::check_id_collision(&extension_id, "extension")?;
-
-    let extension_dir = paths::extension(&extension_id)?;
-    if extension_dir.exists() {
-        return Err(Error::validation_invalid_argument(
-            "extension_id",
-            format!("Extension {} already exists", extension_id),
-            Some(extension_id),
-            None,
-        ));
-    }
-
-    local_files::ensure_app_dirs()?;
-
-    // Clone to a temp directory first so we can detect monorepos before
-    // committing to the final extension location.
-    let extensions_dir = paths::extensions()?;
-    let temp_dir = extensions_dir.join(format!(".clone-tmp-{}", extension_id));
-    if temp_dir.exists() {
-        std::fs::remove_dir_all(&temp_dir).map_err(|e| {
-            Error::internal_io(e.to_string(), Some("clean stale temp dir".to_string()))
-        })?;
-    }
-
-    git::clone_repo_at_ref(url, &temp_dir, revision)?;
-
-    // Capture source revision before resolve_cloned_extension may discard .git
-    // (monorepo installs extract only the subdirectory, losing git history).
-    let source_revision = git::short_head_revision(&temp_dir);
-
-    // Determine what was cloned and install accordingly.
-    let result = resolve_cloned_extension(&temp_dir, &extension_id, &extension_dir, url);
-
-    // Always clean up the temp clone dir (may already be renamed on success).
-    if temp_dir.exists() {
-        let _ = std::fs::remove_dir_all(&temp_dir);
-    }
-
-    let extension_id = result?;
-
-    // Write source metadata so it survives even when .git is discarded.
-    if let Some(ref rev) = source_revision {
-        let _ = std::fs::write(extension_dir.join(".source-revision"), rev);
-    }
-    let _ = std::fs::write(extension_dir.join(".source-url"), url);
-
-    // Auto-run setup if extension defines a setup_command
-    // Setup is best-effort: install succeeds even if setup fails
-    if let Ok(extension) = load_extension(&extension_id) {
-        if extension
-            .runtime()
-            .is_some_and(|r| r.setup_command.is_some())
-        {
-            let _ = run_setup(&extension_id);
-        }
-    }
-
-    Ok(InstallResult {
-        extension_id,
-        url: url.to_string(),
-        path: extension_dir,
-        source_revision,
-    })
-}
-
-/// After cloning a repo to a temp dir, figure out whether it's a single-extension
-/// repo or a monorepo and move the right content to the final extension directory.
-///
-/// Returns the installed extension ID on success.
-pub(crate) fn resolve_cloned_extension(
-    temp_dir: &Path,
-    extension_id: &str,
-    extension_dir: &Path,
-    _url: &str,
-) -> Result<String> {
-    let manifest_at_root = temp_dir.join(format!("{}.json", extension_id));
-
-    // Case 1: Single-extension repo — manifest at clone root.
-    if manifest_at_root.exists() {
-        std::fs::rename(temp_dir, extension_dir).map_err(|e| {
-            Error::internal_io(e.to_string(), Some("move cloned extension".to_string()))
-        })?;
-        return Ok(extension_id.to_string());
-    }
-
-    // Case 2: Monorepo — target extension exists as a subdirectory.
-    let subdir = temp_dir.join(extension_id);
-    let manifest_in_subdir = subdir.join(format!("{}.json", extension_id));
-
-    if subdir.is_dir() && manifest_in_subdir.exists() {
-        // Validate the manifest is parseable before moving.
-        let content = local_files::local().read(&manifest_in_subdir)?;
-        let _manifest: ExtensionManifest = from_str(&content)?;
-
-        install_shared_assets_from_root(temp_dir, extension_dir)?;
-
-        // Move just the subdirectory to the final extension location.
-        rename_dir(&subdir, extension_dir)?;
-        return Ok(extension_id.to_string());
-    }
-
-    // Case 3: No matching extension found. Scan for available extensions to help the user.
-    let available = scan_available_extensions(temp_dir);
-
-    if available.is_empty() {
-        return Err(Error::validation_invalid_argument(
-            "source",
-            format!(
-                "No extension manifest '{}.json' found in cloned repository",
-                extension_id
-            ),
-            None,
-            None,
-        ));
-    }
-
-    let list = available.join(", ");
-    Err(Error::validation_invalid_argument(
-        "id",
-        format!(
-            "Extension '{}' not found in repository. Available extensions: {}",
-            extension_id, list
-        ),
-        Some(extension_id.to_string()),
-        None,
-    )
-    .with_hint(format!(
-        "Install a specific extension with: homeboy extension install <url> --id <extension>\nAvailable: {}",
-        list
-    )))
-}
-
-fn install_shared_assets_from_root(source_root: &Path, extension_dir: &Path) -> Result<()> {
-    let Some(extensions_dir) = extension_dir.parent() else {
-        return Ok(());
-    };
-
-    for shared_dir in ["scripts", "agent-runtimes", "runtime-agent-ci"] {
-        let source = source_root.join(shared_dir);
-        if !source.is_dir() {
-            continue;
-        }
-
-        let target = match shared_dir {
-            "agent-runtimes" => paths::agent_runtimes()?,
-            "runtime-agent-ci" => paths::homeboy()?.join(shared_dir),
-            _ => extensions_dir.join(shared_dir),
-        };
-        if target.exists() {
-            std::fs::remove_dir_all(&target).map_err(|e| {
-                Error::internal_io(
-                    e.to_string(),
-                    Some(format!("replace shared extension {shared_dir}")),
-                )
-            })?;
-        }
-        copy_dir_recursive(&source, &target)?;
-    }
-
-    Ok(())
-}
-
-pub(crate) fn install_linked_shared_assets(
-    source: &Path,
-    extension_dir: &Path,
-    source_root: Option<&Path>,
-) -> Result<()> {
-    if let Some(source_root) = source_root {
-        return install_shared_assets_from_root(source_root, extension_dir);
-    }
-
-    if let Some(parent) = source.parent() {
-        install_shared_assets_from_root(parent, extension_dir)?;
-    }
-    Ok(())
-}
-
-/// Scan a cloned repo for subdirectories that contain a matching manifest file.
-/// Returns a sorted list of extension IDs found.
-fn scan_available_extensions(repo_dir: &Path) -> Vec<String> {
-    let mut found = Vec::new();
-    if let Ok(entries) = std::fs::read_dir(repo_dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path.is_dir() {
-                if let Some(dir_name) = path.file_name().and_then(|n| n.to_str()) {
-                    // Skip hidden dirs (.git, .github, etc.)
-                    if dir_name.starts_with('.') {
-                        continue;
-                    }
-                    let manifest = path.join(format!("{}.json", dir_name));
-                    if manifest.exists() {
-                        found.push(dir_name.to_string());
-                    }
-                }
-            }
-        }
-    }
-    found.sort();
-    found
-}
-
-/// Move a directory, falling back to recursive copy + delete if rename fails
-/// (e.g., across filesystem boundaries).
-pub(crate) fn rename_dir(from: &Path, to: &Path) -> Result<()> {
-    if std::fs::rename(from, to).is_ok() {
-        return Ok(());
-    }
-
-    // Fallback: recursive copy then remove source.
-    copy_dir_recursive(from, to)?;
-    std::fs::remove_dir_all(from)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("remove source after copy".into())))?;
-    Ok(())
-}
-
-/// Recursively copy a directory tree.
-///
-/// Thin wrapper over [`crate::core::io::copy_tree`] with the legacy
-/// extension-lifecycle entry policy (copy any non-directory entry).
-fn copy_dir_recursive(src: &Path, dst: &Path) -> Result<()> {
-    crate::core::io::copy_tree(
-        src,
-        dst,
-        "extension.lifecycle.copy_dir_recursive",
-        crate::core::io::EntryPolicy::CopyAnyNonDir,
-    )
-}
-
-/// Install a extension by symlinking a local directory.
-fn install_from_path(
-    source_path: &str,
-    id_override: Option<&str>,
-    source_root: Option<&Path>,
-) -> Result<InstallResult> {
-    let source = Path::new(source_path);
-
-    // Resolve to absolute path
-    let source = if source.is_absolute() {
-        source.to_path_buf()
-    } else {
-        std::env::current_dir()
-            .map_err(|e| Error::internal_io(e.to_string(), Some("get current dir".to_string())))?
-            .join(source)
-    };
-
-    if !source.exists() {
-        return Err(Error::validation_invalid_argument(
-            "source",
-            format!("Path does not exist: {}", source.display()),
-            Some(source_path.to_string()),
-            None,
-        ));
-    }
-
-    // Derive extension ID from directory name or override
-    let dir_name = source.file_name().and_then(|n| n.to_str()).ok_or_else(|| {
-        Error::validation_invalid_argument(
-            "source",
-            "Could not determine directory name",
-            Some(source_path.to_string()),
-            None,
-        )
-    })?;
-
-    let extension_id = match id_override {
-        Some(id) => slugify_id(id)?,
-        None => slugify_id(dir_name)?,
-    };
-
-    // Check cross-entity name collision before checking extension-specific existence
-    config::check_id_collision(&extension_id, "extension")?;
-
-    let manifest_path = manifest_path_for_extension(&source, &extension_id);
-    if !manifest_path.exists() {
-        if id_override.is_some() {
-            let monorepo_extension = source.join(&extension_id);
-            let monorepo_manifest = manifest_path_for_extension(&monorepo_extension, &extension_id);
-            if monorepo_manifest.exists() {
-                return install_from_path(
-                    &monorepo_extension.to_string_lossy(),
-                    Some(&extension_id),
-                    Some(&source),
-                );
-            }
-        }
-
-        return Err(Error::validation_invalid_argument(
-            "source",
-            format!("No {}.json found at {}", extension_id, source.display()),
-            Some(source_path.to_string()),
-            None,
-        ));
-    }
-
-    // Validate manifest is parseable
-    let manifest_content = local_files::local().read(&manifest_path)?;
-    let _manifest: ExtensionManifest = from_str(&manifest_content)?;
-
-    let extension_dir = paths::extension(&extension_id)?;
-    if extension_dir.exists() {
-        return Err(Error::validation_invalid_argument(
-            "extension_id",
-            format!(
-                "Extension '{}' already exists at {}",
-                extension_id,
-                extension_dir.display()
-            ),
-            Some(extension_id),
-            None,
-        ));
-    }
-
-    local_files::ensure_app_dirs()?;
-
-    install_linked_shared_assets(&source, &extension_dir, source_root)?;
-
-    // Create symlink
-    #[cfg(unix)]
-    std::os::unix::fs::symlink(&source, &extension_dir)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("create symlink".to_string())))?;
-
-    #[cfg(windows)]
-    std::os::windows::fs::symlink_dir(&source, &extension_dir)
-        .map_err(|e| Error::internal_io(e.to_string(), Some("create symlink".to_string())))?;
-
-    // For linked (local) extensions, read revision from the source dir if it's a git repo
-    let source_revision = git::short_head_revision(&source);
-
-    Ok(InstallResult {
-        extension_id,
-        url: source.to_string_lossy().to_string(),
-        path: extension_dir,
-        source_revision,
-    })
-}
-
-/// Update an installed extension by pulling latest changes.
-pub fn update(extension_id: &str, force: bool) -> Result<UpdateResult> {
-    let extension_dir = paths::extension(extension_id)?;
-    if !extension_dir.exists() {
-        return Err(Error::extension_not_found(extension_id.to_string(), vec![]));
-    }
-
-    // Linked extensions: resolve the symlink target and pull the source repo.
-    // The target may be a subdirectory of a larger repo (e.g. an extensions
-    // monorepo/<extension-id>), so we find the git root and pull from there.
-    if is_extension_linked(extension_id) {
-        return update_linked_extension(extension_id, &extension_dir, force);
-    }
-
-    if !force && !is_extension_update_workdir_clean(&extension_dir, &extension_dir) {
-        return Err(Error::validation_invalid_argument(
-            "extension_id",
-            "Extension has uncommitted changes; update may overwrite them. Use --force to proceed.",
-            Some(extension_id.to_string()),
-            None,
-        ));
-    }
-
-    let source = source_metadata::resolve_source_url(extension_id)?;
-    let source_url = source.url;
-    let mut source_repair = source.repair;
-
-    if extension_dir.join(".git").exists() {
-        git::pull_repo(&extension_dir)?;
-
-        // Update source metadata after pull so it stays current.
-        write_source_metadata(
-            &extension_dir,
-            &source_url,
-            git::short_head_revision(&extension_dir),
-        );
-
-        run_setup_if_configured(extension_id);
-
-        return Ok(UpdateResult {
-            extension_id: extension_id.to_string(),
-            url: source_url,
-            path: extension_dir,
-            linked: false,
-            source_path: None,
-            git_root: None,
-            source_update: ExtensionSourceUpdate::default(),
-            repaired_source_metadata: source_repair.take(),
-        });
-    }
-
-    update_extracted_extension(extension_id, &extension_dir, &source_url)?;
-
-    run_setup_if_configured(extension_id);
-
-    Ok(UpdateResult {
-        extension_id: extension_id.to_string(),
-        url: source_url,
-        path: extension_dir,
-        linked: false,
-        source_path: None,
-        git_root: None,
-        source_update: ExtensionSourceUpdate::default(),
-        repaired_source_metadata: source_repair,
-    })
-}
-
-fn update_extracted_extension(
-    extension_id: &str,
-    extension_dir: &Path,
-    source_url: &str,
-) -> Result<()> {
-    let extensions_dir = paths::extensions()?;
-    let clone_dir = extensions_dir.join(format!(".update-clone-tmp-{}", extension_id));
-    let staged_dir = extensions_dir.join(format!(".update-stage-tmp-{}", extension_id));
-    let backup_dir = extensions_dir.join(format!(".update-backup-tmp-{}", extension_id));
-
-    for stale in [&clone_dir, &staged_dir, &backup_dir] {
-        if stale.exists() {
-            std::fs::remove_dir_all(stale).map_err(|e| {
-                Error::internal_io(
-                    e.to_string(),
-                    Some("clean stale extension update dir".to_string()),
-                )
-            })?;
-        }
-    }
-
-    git::clone_repo(source_url, &clone_dir)?;
-    let source_revision = git::short_head_revision(&clone_dir);
-
-    let result = resolve_cloned_extension(&clone_dir, extension_id, &staged_dir, source_url);
-    if clone_dir.exists() {
-        let _ = std::fs::remove_dir_all(&clone_dir);
-    }
-    result?;
-
-    write_source_metadata(&staged_dir, source_url, source_revision);
-
-    rename_dir(extension_dir, &backup_dir)?;
-    if let Err(err) = rename_dir(&staged_dir, extension_dir) {
-        let _ = rename_dir(&backup_dir, extension_dir);
-        return Err(err);
-    }
-
-    if backup_dir.exists() {
-        let _ = std::fs::remove_dir_all(&backup_dir);
-    }
-
-    Ok(())
-}
-
-pub(crate) fn write_source_metadata(
-    extension_dir: &Path,
-    source_url: &str,
-    source_revision: Option<String>,
-) {
-    if let Some(rev) = source_revision {
-        let _ = std::fs::write(extension_dir.join(".source-revision"), rev);
-    }
-    let _ = std::fs::write(extension_dir.join(".source-url"), source_url);
-}
-
-fn is_extension_update_workdir_clean(git_root: &Path, extension_dir: &Path) -> bool {
-    if !git::is_git_repo(&git_root.to_string_lossy()) {
-        return true;
-    }
-
-    let Ok(output) = Command::new("git")
-        .args(["status", "--porcelain=v1"])
-        .current_dir(git_root)
-        .output()
-    else {
-        return false;
-    };
-    if !output.status.success() {
-        return false;
-    }
-
-    let extension_rel = extension_dir
-        .strip_prefix(git_root)
-        .ok()
-        .map(|path| path.to_string_lossy().replace('\\', "/"));
-    let status = String::from_utf8_lossy(&output.stdout);
-
-    status.lines().all(|line| {
-        dirty_path_from_status_line(line).is_some_and(|path| {
-            is_generated_extension_metadata_path(&path, extension_rel.as_deref())
-        })
-    })
-}
-
-fn dirty_path_from_status_line(line: &str) -> Option<String> {
-    let path = line.get(3..)?.trim();
-    let path = path
-        .rsplit_once(" -> ")
-        .map(|(_, new_path)| new_path)
-        .unwrap_or(path);
-    Some(path.trim_matches('"').replace('\\', "/"))
-}
-
-fn is_generated_extension_metadata_path(path: &str, extension_rel: Option<&str>) -> bool {
-    [".source-url", ".source-revision"].iter().any(|name| {
-        path == *name
-            || extension_rel
-                .filter(|rel| !rel.is_empty())
-                .is_some_and(|rel| path == format!("{rel}/{name}"))
-    })
-}
-
-pub(crate) fn run_setup_if_configured(extension_id: &str) {
-    if let Ok(extension) = load_extension(extension_id) {
-        if extension
-            .runtime()
-            .is_some_and(|r| r.setup_command.is_some())
-        {
-            let _ = run_setup(extension_id);
-        }
-    }
-}
-
-fn update_linked_extension(
-    extension_id: &str,
-    extension_dir: &Path,
-    force: bool,
-) -> Result<UpdateResult> {
-    let source_dir = std::fs::read_link(extension_dir).map_err(|e| {
-        Error::internal_io(
-            e.to_string(),
-            Some(format!("read symlink for {}", extension_id)),
-        )
-    })?;
-    let source_dir = if source_dir.is_absolute() {
-        source_dir
-    } else {
-        extension_dir
-            .parent()
-            .unwrap_or_else(|| Path::new("."))
-            .join(source_dir)
-    };
-    let source_dir = source_dir.canonicalize().unwrap_or(source_dir);
-    let git_root_str = git::get_git_root(&source_dir.to_string_lossy())?;
-    let git_root = PathBuf::from(&git_root_str)
-        .canonicalize()
-        .unwrap_or_else(|_| PathBuf::from(git_root_str));
-    let old_branch = git::current_branch(&git_root);
-    let old_source_revision = git::short_head_revision(&git_root);
-
-    static UPDATED_ROOTS: OnceLock<Mutex<HashMap<PathBuf, Option<String>>>> = OnceLock::new();
-    let updated_roots = UPDATED_ROOTS.get_or_init(|| Mutex::new(HashMap::new()));
-    let cached = updated_roots
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .get(&git_root)
-        .cloned();
-    match cached {
-        Some(None) => {}
-        Some(Some(message)) => return Err(Error::validation_invalid_argument(
-            "extension_id",
-            format!("Linked extension '{}' skipped because shared source repo update previously failed: {}", extension_id, message),
-            Some(extension_id.to_string()),
-            None,
-        )),
-        None => {
-            let result = (|| {
-                if !force && !is_extension_update_workdir_clean(&git_root, &source_dir) {
-                    return Err(Error::validation_invalid_argument(
-                        "extension_id",
-                        format!(
-                            "Linked extension source repo has uncommitted changes for {}. Use --force to proceed.",
-                            extension_id,
-                        ),
-                        Some(extension_id.to_string()),
-                        None,
-                    ));
-                }
-
-                git::update_to_remote_default_branch(&git_root)?;
-
-                Ok(())
-            })();
-            updated_roots
-                .lock()
-                .unwrap_or_else(|e| e.into_inner())
-                .insert(
-                    git_root.clone(),
-                    result.as_ref().err().map(|e| e.message.clone()),
-                );
-            result?;
-        }
-    };
-    install_linked_shared_assets(&source_dir, extension_dir, None)?;
-    run_setup_if_configured(extension_id);
-    let url = format!("linked:{}", source_dir.display());
-    let new_branch = git::current_branch(&git_root);
-    let new_source_revision = git::short_head_revision(&git_root);
-    Ok(UpdateResult {
-        extension_id: extension_id.to_string(),
-        url,
-        path: source_dir.clone(),
-        linked: true,
-        source_path: Some(source_dir.clone()),
-        git_root: Some(git_root),
-        source_update: ExtensionSourceUpdate {
-            old_source_revision,
-            new_source_revision,
-            old_branch,
-            new_branch,
-            update_note: Some(
-                "Linked extension source updated in place; clean linked repos switch to the remote default branch before pulling.".to_string(),
-            ),
-        },
-        repaired_source_metadata: None,
-    })
-}
+mod install_sources;
+use install_sources::{install_configured_extension, install_from_path, install_from_url};
+pub(crate) use install_sources::{
+    install_linked_shared_assets, rename_dir, resolve_cloned_extension,
+};
+
+mod update;
+#[cfg(test)]
+use update::is_extension_update_workdir_clean;
+pub(crate) use update::write_source_metadata;
+pub use update::{check_update_available, read_source_revision, update, UpdateAvailable};
 
 /// Uninstall a extension. Automatically detects symlinks vs cloned directories.
 /// - Symlinked extensions: removes symlink only (source preserved)
@@ -816,84 +169,6 @@ pub fn uninstall(extension_id: &str) -> Result<PathBuf> {
     }
 
     Ok(extension_dir)
-}
-
-/// Check if a git-cloned extension has updates available.
-/// Runs `git fetch` then checks if HEAD is behind the remote tracking branch.
-/// Returns None for linked extensions or if check fails.
-pub fn check_update_available(extension_id: &str) -> Option<UpdateAvailable> {
-    let extension_dir = paths::extension(extension_id).ok()?;
-    if !extension_dir.exists() || is_extension_linked(extension_id) {
-        return None;
-    }
-
-    // Check it's a git repo
-    if !extension_dir.join(".git").exists() {
-        return None;
-    }
-
-    // Fetch latest (best-effort, short timeout)
-    Command::new("git")
-        .args(["fetch", "--quiet"])
-        .current_dir(&extension_dir)
-        .stdin(std::process::Stdio::null())
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .status()
-        .ok()?;
-
-    // Check how many commits we're behind
-    let output = Command::new("git")
-        .args(["rev-list", "HEAD..@{u}", "--count"])
-        .current_dir(&extension_dir)
-        .stdin(std::process::Stdio::null())
-        .output()
-        .ok()?;
-
-    let count_str = String::from_utf8_lossy(&output.stdout).trim().to_string();
-    let behind_count: usize = count_str.parse().ok()?;
-
-    if behind_count == 0 {
-        return None;
-    }
-
-    // Get installed version
-    let extension = load_extension(extension_id).ok()?;
-    let installed_version = extension.version.clone();
-
-    Some(UpdateAvailable {
-        extension_id: extension_id.to_string(),
-        installed_version,
-        behind_count,
-    })
-}
-
-#[derive(Debug, Clone)]
-pub struct UpdateAvailable {
-    pub extension_id: String,
-    pub installed_version: String,
-    pub behind_count: usize,
-}
-
-/// Read the source revision for an installed extension.
-/// Checks (in order): .git directory (git rev-parse), then .source-revision file.
-pub fn read_source_revision(extension_id: &str) -> Option<String> {
-    let extension_dir = paths::extension(extension_id).ok()?;
-    if !extension_dir.exists() {
-        return None;
-    }
-
-    // Try .git first (single-extension repos and linked extensions)
-    if let Some(rev) = git::short_head_revision(&extension_dir) {
-        return Some(rev);
-    }
-
-    // Fall back to .source-revision file (monorepo installs)
-    let rev_file = extension_dir.join(".source-revision");
-    std::fs::read_to_string(&rev_file)
-        .ok()
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
 }
 
 #[cfg(test)]
