+++ src/core/runner/rig_materialization/mod.rs
@@ -7,10 +7,16 @@ use crate::core::{Error, Result};
 use crate::extensions::deps_provider;
 
 use super::{
-    exec, load, materialize_git_dependency, sync_workspace,
+    exec, load, materialize_git_dependency, preflight_remote_argv_path_translation, sync_workspace,
     workspace::{parent_remote_path, sanitize_path_segment},
     RunnerExecOptions, RunnerGitDependencyMaterializationOptions,
     RunnerGitDependencyMaterializationOutput, RunnerWorkspaceSyncMode, RunnerWorkspaceSyncOptions,
+};
+
+mod rig_source_install;
+use rig_source_install::{
+    remote_package_path, remove_runner_installed_rig_source, rig_install_capability_preflight,
+    validate_installed_rig_source,
 };
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -103,26 +109,47 @@ pub(super) fn sync_lab_offload_rigs(
         let removed_source =
             remove_runner_installed_rig_source(runner_id, command_path, remote_cwd, rig_id)?;
 
+        let install_command = vec![
+            command_path.to_string(),
+            "rig".to_string(),
+            "install".to_string(),
+            source.clone(),
+            "--id".to_string(),
+            rig_id.clone(),
+        ];
+
+        // Path-translation preflight: the rig install source has already been
+        // resolved to a runner-side remote path (primary snapshot remote path or
+        // a materialized installed-source path). Assert that no controller-local
+        // primary source path survived un-translated into the dispatched argv
+        // before we hand it to the remote runner, so a missed remap fails loudly
+        // here instead of installing from a non-existent local path (#5285).
+        preflight_remote_argv_path_translation(
+            "Runner rig materialization",
+            runner_id,
+            &install_command,
+            Path::new(primary_local_path),
+            remote_cwd,
+        )?;
+
         let (output, exit_code) = exec(
             runner_id,
             RunnerExecOptions {
                 cwd: Some(remote_cwd.to_string()),
                 project_id: None,
                 allow_diagnostic_ssh: false,
-                command: vec![
-                    command_path.to_string(),
-                    "rig".to_string(),
-                    "install".to_string(),
-                    source.clone(),
-                    "--id".to_string(),
-                    rig_id.clone(),
-                ],
+                command: install_command,
                 env: HashMap::new(),
                 secret_env_names: Vec::new(),
                 capture_patch: false,
                 raw_exec: false,
                 source_snapshot: None,
-                capability_preflight: None,
+                // Validate remote capability parity before dispatch: the install
+                // is executed by the runner-side `homeboy` binary, so require
+                // that tool on the runner. `exec` no-ops this gate for local and
+                // already-capable SSH runners, so behavior is unchanged on a
+                // correctly provisioned runner and fails early otherwise (#5285).
+                capability_preflight: Some(rig_install_capability_preflight()),
                 required_extensions: Vec::new(),
                 require_paths: Vec::new(),
             },
@@ -151,138 +178,6 @@ pub(super) fn sync_lab_offload_rigs(
     }
 
     Ok(synced_rigs)
-}
-
-fn remote_package_path(source_root: &str, package_path: &str, remote_source_root: &str) -> String {
-    let source_root = Path::new(source_root);
-    let package_path = Path::new(package_path);
-    match package_path.strip_prefix(source_root) {
-        Ok(relative) if !relative.as_os_str().is_empty() => Path::new(remote_source_root)
-            .join(relative)
-            .to_string_lossy()
-            .to_string(),
-        _ => remote_source_root.to_string(),
-    }
-}
-
-fn remove_runner_installed_rig_source(
-    runner_id: &str,
-    command_path: &str,
-    remote_cwd: &str,
-    rig_id: &str,
-) -> Result<Option<String>> {
-    let (list_output, list_exit_code) = exec(
-        runner_id,
-        RunnerExecOptions {
-            cwd: Some(remote_cwd.to_string()),
-            project_id: None,
-            allow_diagnostic_ssh: false,
-            command: vec![
-                command_path.to_string(),
-                "rig".to_string(),
-                "sources".to_string(),
-                "list".to_string(),
-            ],
-            env: HashMap::new(),
-            secret_env_names: Vec::new(),
-            capture_patch: false,
-            raw_exec: false,
-            source_snapshot: None,
-            capability_preflight: None,
-            required_extensions: Vec::new(),
-            require_paths: Vec::new(),
-        },
-    )?;
-    if list_exit_code != 0 {
-        return Err(Error::validation_invalid_argument(
-            "rig",
-            format!(
-                "runner dispatch could not inspect installed rig sources on runner `{runner_id}`"
-            ),
-            Some(rig_id.to_string()),
-            Some(vec![list_output.stderr.trim().to_string()]),
-        ));
-    }
-
-    let Some(selector) = installed_source_selector_for_rig(&list_output.stdout, rig_id)? else {
-        return Ok(None);
-    };
-
-    let (remove_output, remove_exit_code) = exec(
-        runner_id,
-        RunnerExecOptions {
-            cwd: Some(remote_cwd.to_string()),
-            project_id: None,
-            allow_diagnostic_ssh: false,
-            command: vec![
-                command_path.to_string(),
-                "rig".to_string(),
-                "sources".to_string(),
-                "remove".to_string(),
-                selector.clone(),
-            ],
-            env: HashMap::new(),
-            secret_env_names: Vec::new(),
-            capture_patch: false,
-            raw_exec: false,
-            source_snapshot: None,
-            capability_preflight: None,
-            required_extensions: Vec::new(),
-            require_paths: Vec::new(),
-        },
-    )?;
-    if remove_exit_code != 0 {
-        return Err(Error::validation_invalid_argument(
-            "rig",
-            format!("runner dispatch could not remove stale rig source for `{rig_id}` on runner `{runner_id}`"),
-            Some(rig_id.to_string()),
-            Some(vec![remove_output.stderr.trim().to_string()]),
-        ));
-    }
-
-    Ok(Some(selector))
-}
-
-fn installed_source_selector_for_rig(stdout: &str, rig_id: &str) -> Result<Option<String>> {
-    let value: serde_json::Value = serde_json::from_str(stdout).map_err(|e| {
-        Error::validation_invalid_json(
-            e,
-            Some("parse runner rig sources list output".to_string()),
-            Some(stdout.chars().take(200).collect()),
-        )
-    })?;
-    let sources = value
-        .get("data")
-        .and_then(|data| data.get("report"))
-        .and_then(|report| report.get("sources"))
-        .and_then(|sources| sources.as_array())
-        .cloned()
-        .unwrap_or_default();
-
-    for source in sources {
-        let has_rig = source
-            .get("rigs")
-            .and_then(|rigs| rigs.as_array())
-            .is_some_and(|rigs| {
-                rigs.iter().any(|rig| {
-                    rig.get("id")
-                        .and_then(|id| id.as_str())
-                        .is_some_and(|id| id == rig_id)
-                })
-            });
-        if !has_rig {
-            continue;
-        }
-        for key in ["package_id", "package_path", "source"] {
-            if let Some(selector) = source.get(key).and_then(|value| value.as_str()) {
-                if !selector.trim().is_empty() {
-                    return Ok(Some(selector.to_string()));
-                }
-            }
-        }
-    }
-
-    Ok(None)
 }
 
 pub(super) fn remap_bench_rig_default_component_to_primary_snapshot(
@@ -423,36 +318,6 @@ pub(super) fn sync_lab_offload_rig_component_dependencies(
     })
 }
 
-fn validate_installed_rig_source(
-    rig_id: &str,
-    source_root: &str,
-    package_path: &str,
-) -> Result<()> {
-    if Path::new(source_root).is_dir() {
-        return Ok(());
-    }
-
-    Err(Error::validation_invalid_argument(
-        "rig",
-        format!(
-            "runner dispatch cannot materialize rig `{rig_id}` because its installed source metadata points at a missing path"
-        ),
-        Some(source_root.to_string()),
-        Some(vec![
-            format!(
-                "Repair the rig source metadata with: homeboy rig install {} --id {} --reinstall",
-                shell_arg(package_path),
-                shell_arg(rig_id)
-            ),
-            format!(
-                "Or remove the stale source metadata with: homeboy rig sources remove {}",
-                shell_arg(package_path)
-            ),
-            "Run `homeboy rig sources list` to inspect installed rig sources.".to_string(),
-        ]),
-    ))
-}
-
 fn prepare_rig_component_dependency(dependency: &RigComponentDependency) -> Result<()> {
     let path = Path::new(&dependency.local_checkout_root);
     let component = Component {
@@ -470,16 +335,6 @@ fn prepare_rig_component_dependency(dependency: &RigComponentDependency) -> Resu
         provider.install(&component, path)?;
     }
     Ok(())
-}
-
-fn shell_arg(value: &str) -> String {
-    if value
-        .chars()
-        .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | '.' | '/' | ':' | '='))
-    {
-        return value.to_string();
-    }
-    format!("'{}'", value.replace('\'', "'\\''"))
 }
 
 /// Join a runner-side checkout root with the component's required subpath.
@@ -808,80 +663,6 @@ fn push_unique<T: PartialEq>(items: &mut Vec<T>, item: T) {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn installed_source_selector_prefers_package_id_for_matching_rig() {
-        let stdout = serde_json::json!({
-            "success": true,
-            "data": {
-                "command": "rig.sources.list",
-                "report": {
-                    "sources": [
-                        {
-                            "source": "/runner/old-source",
-                            "package_id": "old-source",
-                            "package_path": "/runner/old-source",
-                            "rigs": [{ "id": "target-rig" }]
-                        }
-                    ]
-                }
-            }
-        })
-        .to_string();
-
-        assert_eq!(
-            installed_source_selector_for_rig(&stdout, "target-rig").unwrap(),
-            Some("old-source".to_string())
-        );
-    }
-
-    #[test]
-    fn installed_source_selector_ignores_unrelated_sources() {
-        let stdout = serde_json::json!({
-            "success": true,
-            "data": {
-                "report": {
-                    "sources": [
-                        {
-                            "source": "/runner/old-source",
-                            "package_id": "old-source",
-                            "package_path": "/runner/old-source",
-                            "rigs": [{ "id": "other-rig" }]
-                        }
-                    ]
-                }
-            }
-        })
-        .to_string();
-
-        assert_eq!(
-            installed_source_selector_for_rig(&stdout, "target-rig").unwrap(),
-            None
-        );
-    }
-
-    #[test]
-    fn stale_installed_rig_source_metadata_suggests_reinstall() {
-        let err = validate_installed_rig_source(
-            "woocommerce-performance",
-            "/missing/homeboy-rigs/woocommerce/woocommerce",
-            "/Users/user/Developer/sample-rigs@issue-323-runtime-fresh/sample-plugin/sample-plugin",
-        )
-        .expect_err("missing source should fail");
-
-        assert!(err.message.contains("installed source metadata"));
-        assert_eq!(err.details["field"], "rig");
-        let hints = err.details["tried"]
-            .as_array()
-            .expect("tried hints")
-            .iter()
-            .filter_map(|hint| hint.as_str())
-            .collect::<Vec<_>>()
-            .join("\n");
-        assert!(hints.contains("homeboy rig install"));
-        assert!(hints.contains("--id woocommerce-performance --reinstall"));
-        assert!(hints.contains("homeboy rig sources list"));
-    }
 
     #[test]
     fn extracts_unique_bench_rig_ids_for_lab_materialization() {
